pyproject.toml
[project]
name = "dl-vsix"
version = "0.1.0"
description = "Download VSIX bundles for offline extension installation"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "sco1", email = "sco1.git@gmail.com"}
]

readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Typing :: Typed",
    "Private :: Do Not Upload",
]

requires-python = ">=3.12"
dependencies = [
    "httpx~=0.28",
    "platformdirs~=4.3",
    "typer-slim~=0.16",
]

[project.urls]
Homepage = "https://github.com/sco1/"
Documentation = "https://github.com/sco1/dl-vsix/blob/main/README.md"
Repository = "https://github.com/sco1/dl-vsix.git"
Issues = "https://github.com/sco1/dl-vsix/issues"
Changelog = "https://github.com/sco1/dl-vsix/blob/main/CHANGELOG.md"

[project.scripts]
dl_vsix = "dl_vsix.cli:dl_vsix_cli"

[tool.uv]
dev-dependencies = [
    "black~=25.0",
    "cogapp~=3.5",
    "flake8~=7.1",
    "flake8-annotations~=3.1",
    "flake8-define-class-attributes~=0.2",
    "isort~=6.0",
    "mypy~=1.14",
    "pre-commit~=4.0",
    "pytest~=8.3",
    "pytest-check~=2.4",
    "pytest-cov~=6.0",
    "pytest-mock~=3.14",
    "pytest-randomly~=3.16",
    "ruff~=0.9",
    "rumdl~=0.0.120",
    "sco1-bumper~=2.0",
    "tox~=4.23",
    "tox-uv~=1.17",
]

[tool.black]
line-length = 100

[tool.isort]
case_sensitive = true
known_first_party = "dl-vsix,tests"
no_lines_before = "LOCALFOLDER"
order_by_type = false
profile = "black"
line_length = 100

[tool.mypy]
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
enable_error_code = "exhaustive-match"
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.rumdl]
line-length = 100
disable = ["MD013", "MD033"]

[tool.uv.build-backend]
module-name = "dl_vsix"
module-root = ""

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"
